apps/old-gp2-browser/prototype/requirements.txt
pandas==2.2.3
protobuf==5.29.5
streamlit==1.41.1
google-cloud-storage
